Updated coherence to 1.3.0

+++ Assets/coherence/baked/Interop/_3889a458e94666d4784df30d8dd06d7d_df8fe4d8a6104b7ab780e439afba5391.cs
@@ -40,7 +40,7 @@ namespace Coherence.Generated
             var comp = (Interop*)data;
             orig.oldEffectID = comp->oldEffectID;
             orig.newEffectID = comp->newEffectID;
-            orig.syncConfigID = comp->syncConfigID.Data != null ? System.Text.Encoding.UTF8.GetString((byte*)comp->syncConfigID.Data, comp->syncConfigID.Length) : null;
+            orig.syncConfigID = comp->syncConfigID.Data != null ? System.Text.Encoding.UTF8.GetString((byte*)comp->syncConfigID.Data, (int)comp->syncConfigID.Length) : null;
             return orig;
         }
