+++ Assets/coherence/baked/Interop/_3889a458e94666d4784df30d8dd06d7d_f8f688eb08cb493facb6d766417377d4.cs
@@ -34,7 +34,7 @@ namespace Coherence.Generated
 
             var orig = new _3889a458e94666d4784df30d8dd06d7d_f8f688eb08cb493facb6d766417377d4();
             var comp = (Interop*)data;
-            orig.syncConfigID = comp->syncConfigID.Data != null ? System.Text.Encoding.UTF8.GetString((byte*)comp->syncConfigID.Data, comp->syncConfigID.Length) : null;
+            orig.syncConfigID = comp->syncConfigID.Data != null ? System.Text.Encoding.UTF8.GetString((byte*)comp->syncConfigID.Data, (int)comp->syncConfigID.Length) : null;
             return orig;
         }
